backend/auth_service/requirements.txt
# Auth Service - Dependencias
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-multipart==0.0.6
slowapi==0.1.9
